src/localguide/pyproject.toml
[project]
name = "localguide"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.23.0",
    "fastapi[standard]>=0.115.12",
    "agent-framework-azure-ai>=1.0.0b251028",
    "uvicorn>=0.34.2",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
]
